Fixes #4118 - support for command testing moved to core
